fix: resolve PostgreSQL float decode and pinned_messages SQL syntax
- Use server.state.db_is_postgres in pin_message test call instead of
  hardcoded false, which sent SQLite INSERT OR IGNORE to PostgreSQL
- Change soundboard volume column from REAL (float4) to DOUBLE PRECISION
  (float8) in PostgreSQL schema to match Rust f64
- Add get_f64 helper (like get_bool) for safe float decoding across backends
- Switch soundboard queries from query_as tuple to manual row extraction

+++ migrations/postgres/001_initial_schema.sql
@@ -308,7 +308,7 @@ CREATE TABLE IF NOT EXISTS soundboard_sounds (
     audio_path TEXT,
     audio_content_type TEXT,
     audio_size INTEGER,
-    volume REAL NOT NULL DEFAULT 1.0,
+    volume DOUBLE PRECISION NOT NULL DEFAULT 1.0,
     creator_id TEXT REFERENCES users(id),
     created_at TEXT NOT NULL DEFAULT (to_char(now() at time zone 'UTC', 'YYYY-MM-DD HH24:MI:SS')),
     updated_at TEXT NOT NULL DEFAULT (to_char(now() at time zone 'UTC', 'YYYY-MM-DD HH24:MI:SS'))
